feat(outlook): remove separate get token step [E-3489]
