+++ apps/outlook/drizzle/0000_far_doctor_doom.sql
@@ -2,7 +2,6 @@ CREATE TABLE IF NOT EXISTS "organisations" (
 	"id" uuid PRIMARY KEY NOT NULL,
 	"tenant_id" text NOT NULL,
 	"region" text NOT NULL,
-	"token" text NOT NULL,
 	"created_at" timestamp DEFAULT now() NOT NULL,
 	"last_sync_started_at" timestamp
 );
