+++ apps/outlook/drizzle/meta/_journal.json
@@ -5,8 +5,8 @@
     {
       "idx": 0,
       "version": "5",
-      "when": 1749191770847,
-      "tag": "0000_polite_scalphunter",
+      "when": 1751277215170,
+      "tag": "0000_far_doctor_doom",
       "breakpoints": false
     }
   ]
