+++ apps/outlook/src/database/schema/organisations.ts
@@ -5,7 +5,6 @@ export const organisationsTable = pgTable('organisations', {
   id: uuid('id').primaryKey(),
   tenantId: text('tenant_id').notNull(),
   region: text('region').notNull(),
-  token: text('token').notNull(),
   createdAt: timestamp('created_at').defaultNow().notNull(),
   lastSyncStartedAt: timestamp('last_sync_started_at'),
 });
