+++ apps/outlook/src/inngest/functions/common/get-organisation.test.ts
@@ -16,25 +16,22 @@ describe('get-organisation', () => {
       {
         id: '4ef9c9ad-947b-4ec2-bbc4-cbe3190eee51',
         tenantId: 'c647a27f-7060-4e8d-acc9-05a42218235b',
-        token: 'token-org-1',
         region: 'eu',
       },
       {
         id: '791186a4-1df0-4c57-a442-ba4ff332b101',
         tenantId: 'f5151789-9d9d-4a94-ac74-7ac456422f26',
-        token: 'token-org-2',
         region: 'us',
       },
     ]);
   });
   test('should return desired organisation fields', async () => {
     const [result] = setup({
-      columns: ['token', 'tenantId', 'region'],
+      columns: ['tenantId', 'region'],
       organisationId: '4ef9c9ad-947b-4ec2-bbc4-cbe3190eee51',
     });
 
     await expect(result).resolves.toStrictEqual({
-      token: 'token-org-1',
       tenantId: 'c647a27f-7060-4e8d-acc9-05a42218235b',
       region: 'eu',
     });
@@ -42,7 +39,7 @@ describe('get-organisation', () => {
 
   test('should throw an error when there is no organisation with a given id ', async () => {
     const [result] = setup({
-      columns: ['token', 'tenantId', 'region'],
+      columns: ['tenantId', 'region'],
       organisationId: '00000000-0000-0000-0000-000000000002',
     });
