+++ apps/outlook/src/inngest/functions/common/index.ts
@@ -2,13 +2,8 @@ import type { GetOrganisationEvents } from './get-organisation';
 import type { RemoveOrganisationEvents } from './remove-organisation';
 import { removeOrganisation } from './remove-organisation';
 import { getOrganisation } from './get-organisation';
-import { getToken } from './get-token';
 import type { OrganisationEvents } from './organisation';
-import { type GetTokenEvents } from './get-token';
 
-export type CommonEvents = GetOrganisationEvents &
-  OrganisationEvents &
-  RemoveOrganisationEvents &
-  GetTokenEvents;
+export type CommonEvents = GetOrganisationEvents & OrganisationEvents & RemoveOrganisationEvents;
 
-export const commonFunctions = [getOrganisation, removeOrganisation, getToken];
+export const commonFunctions = [getOrganisation, removeOrganisation];
