+++ apps/outlook/src/inngest/functions/common/remove-organisation.test.ts
@@ -18,13 +18,11 @@ describe('remove-organisation', () => {
       {
         id: '4ef9c9ad-947b-4ec2-bbc4-cbe3190eee51',
         tenantId: 'c647a27f-7060-4e8d-acc9-05a42218235b',
-        token: 'token-org-1',
         region: 'eu',
       },
       {
         id: '791186a4-1df0-4c57-a442-ba4ff332b101',
         tenantId: 'f5151789-9d9d-4a94-ac74-7ac456422f26',
-        token: 'token-org-2',
         region: 'us',
       },
     ]);
@@ -41,7 +39,6 @@ describe('remove-organisation', () => {
         id: '791186a4-1df0-4c57-a442-ba4ff332b101',
         tenantId: 'f5151789-9d9d-4a94-ac74-7ac456422f26',
         lastSyncStartedAt: null,
-        token: 'token-org-2',
         region: 'us',
       },
     ]);
