+++ apps/outlook/src/inngest/functions/index.ts
@@ -1,19 +1,13 @@
 import { type CommonEvents, commonFunctions } from './common';
 import { type UsersEvents, usersFunctions } from './users';
-import { type TokenEvents, tokenFunctions } from './tokens';
 import { type ThirdPartyAppsEvents, thirdPartyAppsFunctions } from './third-party-apps';
 import { type MicrosoftEvents, microsoftFunctions } from './microsoft';
 
 export const inngestFunctions = [
   ...commonFunctions,
   ...usersFunctions,
-  ...tokenFunctions,
   ...thirdPartyAppsFunctions,
   ...microsoftFunctions,
 ];
 
-export type InngestEvents = CommonEvents &
-  UsersEvents &
-  TokenEvents &
-  ThirdPartyAppsEvents &
-  MicrosoftEvents;
+export type InngestEvents = CommonEvents & UsersEvents & ThirdPartyAppsEvents & MicrosoftEvents;
