--- apps/outlook/src/inngest/functions/tokens/index.ts
@@ -1,6 +1,0 @@
-import { refreshToken, type RefreshTokenEvents } from './refresh-token';
-import { scheduleTokenRefresh } from './schedule-token-refresh';
-
-export type TokenEvents = RefreshTokenEvents;
-
-export const tokenFunctions = [refreshToken, scheduleTokenRefresh];
